Move pkg/gen to use keep-common promise generator
This rips out all code generation code, as well as moving references to
generated promises to use the naming convention that the new independent
promise generator uses.

It also adds a keep-common dependency to our Gopkg.toml, requiring the
relevant generator tool package, since it's not referenced directly from
Go code. The vendored version of that package is referenced in the
go:generate clause and run using `go run`.
